% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_firth_loglik.R
\name{compute_firth_loglik}
\alias{compute_firth_loglik}
\title{Compute Firth penalized log likelihood
Compute the Firth penalized log likelihood for given values of B and z.}
\usage{
compute_firth_loglik(Y, X, B, z, X_tilde, W, X_tilde_trans)
}
\arguments{
\item{Y}{An outcome matrix with n rows (for samples) and J columns (for KOs) containing coverage data.}

\item{X}{Design matrix with n rows (for samples) and p columns (for covariates).}

\item{B}{Beta parameter matrix with p rows (for covariates) and J columns (for KOs).}

\item{z}{z parameter vector of length n.}

\item{X_tilde}{An expanded version of the design matrix.}

\item{W}{A diagonal matrix with expected values of Y tilde on the diagonal.}
}
\value{
The value of the log likelihood.
}
\description{
Compute Firth penalized log likelihood
Compute the Firth penalized log likelihood for given values of B and z.
}
