% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_info.R
\name{compute_info}
\alias{compute_info}
\title{Compute information
Calculate the information matrix with MLEs of parameters.}
\usage{
compute_info(X, B, z, constraint, constraint_cat)
}
\arguments{
\item{X}{Design matrix with n rows (for samples) and p columns (for covariates), should have a leading intercept column of \code{1}s, can be replaced by \code{formula_rhs} and \code{covariate_data}.}

\item{B}{B parameter matrix.}

\item{z}{z parameter vector.}

\item{constraint}{Should be "scc" for single category constraint or "mc" for mean constraint.}

\item{constraint_cat}{Category to constrain coefficients (for single category the B parameters for this
category will equal 0, for mean the B parameters for this category will equal the negative sum of the others.}
}
\value{
The information matrix for a constrained likelihood.
}
\description{
Compute information
Calculate the information matrix with MLEs of parameters.
}
