% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_info_cstr.R
\name{compute_info_cstr}
\alias{compute_info_cstr}
\title{Compute information matrix with constraint
Compute the information matrix for all parameters for given values of B and z with constraint.}
\usage{
compute_info_cstr(X, B, z, constraint, constraint_cat = 1)
}
\arguments{
\item{X}{Design matrix with n rows (for samples) and p columns (for covariates).}

\item{B}{Beta parameter matrix with p rows (for covariates) and J columns (for categories).}

\item{z}{z parameter vector of length n.}

\item{constraint}{Should be "scc" for single category constraint, "mc" for mean constraint, or "msc" for mean
over a subset constraint.}

\item{constraint_cat}{Category to constrain coefficients to equal the negative sum of all other categories.}
}
\value{
The square \code{p*J + n} information matrix, evaluated at the values of B and z with the constraint.
}
\description{
Compute information matrix with constraint
Compute the information matrix for all parameters for given values of B and z with constraint.
}
