% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_J.R
\name{compute_J}
\alias{compute_J}
\title{Compute the Jacobian for Newton-Raphson for updating B vectors.}
\usage{
compute_J(X, B, z, j)
}
\arguments{
\item{X}{Design matrix with n rows (for samples) and p columns (for covariates).}

\item{B}{Beta parameter matrix with p rows (for covariates) and J columns (for KOs).}

\item{z}{z parameter vector of length n.}

\item{j}{Category being updated.}
}
\value{
\code{k} by \code{k} matrix J.
}
\description{
Compute the Jacobian for Newton-Raphson for updating B vectors.
}
