% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_loglik.R
\name{compute_loglik}
\alias{compute_loglik}
\title{Compute log likelihood
Compute the log likelihood for given values of B and z.}
\usage{
compute_loglik(Y, X, B, z)
}
\arguments{
\item{Y}{An outcome matrix with n rows (for samples) and J columns (for KOs) containing coverage data.}

\item{X}{Design matrix with n rows (for samples) and p columns (for covariates).}

\item{B}{Beta parameter matrix with p rows (for covariates) and J columns (for KOs).}

\item{z}{z parameter vector of length n.}
}
\value{
The value of the log likelihood.
}
\description{
Compute log likelihood
Compute the log likelihood for given values of B and z.
}
