% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_loglik_constrained.R
\name{compute_loglik_constrained}
\alias{compute_loglik_constrained}
\title{Compute log likelihood with constraints
Compute the log likelihood for given values of B and z and constraints.}
\usage{
compute_loglik_constrained(
  Y,
  X,
  B,
  z,
  lambda,
  lambda_null,
  constraint_fn,
  null_j,
  null_k
)
}
\arguments{
\item{Y}{An outcome matrix with n rows (for samples) and J columns (for KOs) containing coverage data.}

\item{X}{Design matrix with n rows (for samples) and p columns (for covariates).}

\item{B}{Beta parameter matrix with p rows (for covariates) and J columns (for KOs).}

\item{z}{z parameter vector of length n.}

\item{lambda}{Vector of Lagrange multipliers for identifiability constraints.}

\item{lambda_null}{Lagrange multiplier for null hypothesis constraint.}

\item{constraint_fn}{A constraint function to make the B matrix identifiable.#'}

\item{null_j}{Category for which covariate \code{k} is set to \code{0} under the null hypothesis.}

\item{null_k}{Coefficient of the covariate in design matrix that is set to \code{0} under the null hypothesis.}
}
\value{
The value of the log likelihood.
}
\description{
Compute log likelihood with constraints
Compute the log likelihood for given values of B and z and constraints.
}
