% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_score_var_cstr.R
\name{compute_score_var_cstr}
\alias{compute_score_var_cstr}
\title{Variance of score
Calculate the empirical variance of the score for the constrained likelihood.}
\usage{
compute_score_var_cstr(Y, X, B, z, constraint, constraint_cat)
}
\arguments{
\item{Y}{An outcome matrix with n rows (for samples) and J columns (for categories) containing abundance data.}

\item{X}{Design matrix with n rows (for samples) and p columns (for covariates).}

\item{B}{Beta parameter matrix with p rows (for covariates) and J columns (for categories).}

\item{z}{z parameter vector of length n.}

\item{constraint}{Should be "scc" for single category constraint, "mc" for mean constraint, or "msc" for mean
over a subset constraint.}

\item{constraint_cat}{Category to constrain coefficients to equal the negative sum of all other categories.}
}
\value{
A matrix giving the empirical variance of the score vector.
}
\description{
Variance of score
Calculate the empirical variance of the score for the constrained likelihood.
}
