% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_scores.R
\name{compute_scores}
\alias{compute_scores}
\title{Compute score vector
Compute the score equations for all parameters for given values of B and z.}
\usage{
compute_scores(X, Y, B, z)
}
\arguments{
\item{X}{Design matrix with n rows (for samples) and p columns (for covariates).}

\item{Y}{An outcome matrix with n rows (for samples) and J columns (for KOs) containing coverage data.}

\item{B}{Beta parameter matrix with p rows (for covariates) and J columns (for KOs).}

\item{z}{z parameter vector of length n.}
}
\value{
The value of each score equation, evaluated at the values of B and z.
}
\description{
Compute score vector
Compute the score equations for all parameters for given values of B and z.
}
