% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_scores_constraint.R
\name{compute_scores_constraint}
\alias{compute_scores_constraint}
\title{Compute score vector from likelihood with Lagrange multipliers
Compute the score equations for all parameters for given values of B and z.}
\usage{
compute_scores_constraint(X, Y, B, z, constraint_fn, null_j, null_k)
}
\arguments{
\item{X}{Design matrix with n rows (for samples) and p columns (for covariates).}

\item{Y}{An outcome matrix with n rows (for samples) and J columns (for KOs) containing coverage data.}

\item{B}{Beta parameter matrix with p rows (for covariates) and J columns (for KOs).}

\item{z}{z parameter vector of length n.}

\item{constraint_fn}{A constraint function to make the B matrix identifiable.#'}

\item{null_j}{Category for which covariate \code{k} is set to \code{0} under the null hypothesis.}

\item{null_k}{Coefficient of the covariate in design matrix that is set to \code{0} under the null hypothesis.}
}
\value{
The value of each score equation, evaluated at the values of B and z.
}
\description{
Compute score vector from likelihood with Lagrange multipliers
Compute the score equations for all parameters for given values of B and z.
}
