% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_scores_cstr.R
\name{compute_scores_cstr}
\alias{compute_scores_cstr}
\title{Compute score vector with constraint
Compute the score equations for all parameters for given values of B and z with constraint.}
\usage{
compute_scores_cstr(X, Y, B, z, constraint, constraint_cat, subset_j)
}
\arguments{
\item{X}{Design matrix with n rows (for samples) and p columns (for covariates).}

\item{Y}{An outcome matrix with n rows (for samples) and J columns (for categories) containing abundance data.}

\item{B}{Beta parameter matrix with p rows (for covariates) and J columns (for categories).}

\item{z}{z parameter vector of length n.}

\item{constraint}{Should be "scc" for single category constraint, "mc" for mean constraint, or "msc" for mean
over a subset constraint.}

\item{constraint_cat}{Category to constrain coefficients to equal the negative sum of all other categories.}

\item{subset_j}{Indices of categories to include in constraint for the mean over a subset constraint.}
}
\value{
The value of each score equation, evaluated at the values of B and z with the constraint.
}
\description{
Compute score vector with constraint
Compute the score equations for all parameters for given values of B and z with constraint.
}
