% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_var_single_score.R
\name{compute_var_single_score}
\alias{compute_var_single_score}
\title{Estimate variance of single score value
Calculate the empirical variance of the score for the parameter tested.}
\usage{
compute_var_single_score(null_ind, upd_ind, p, D = NULL, info, robust = TRUE)
}
\arguments{
\item{null_ind}{Index of parameter vector corresponding with parameter being tested.}

\item{upd_ind}{Indices of parameter vector, not including parameters from constraint row or parameter implicated in null hypothesis.}

\item{p}{Number of columns in design matrix.}

\item{D}{Estimated empirical variance of score.}

\item{info}{Information matrix.}

\item{robust}{If TRUE uses the D matrix and creates a robust test, if FALSE uses only the information matrix for a model-based test.}
}
\value{
The empirical variance of the score for a single value.
}
\description{
Estimate variance of single score value
Calculate the empirical variance of the score for the parameter tested.
}
