% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_alg1.R
\name{fit_alg1}
\alias{fit_alg1}
\title{Estimates B and z parameters
Estimates B and z parameters using Algorithm 1.}
\usage{
fit_alg1(
  formula_rhs = NULL,
  Y,
  X = NULL,
  covariate_data = NULL,
  B = NULL,
  tolerance = 0.1,
  maxit = 100,
  constraint_fn,
  maxit_glm = 100
)
}
\description{
Estimates B and z parameters
Estimates B and z parameters using Algorithm 1.
}
\examples{
set.seed(4323)
X <- cbind(1,rep(c(0,1),each = 20))
z <- rnorm(40) + 8
b0 <- rnorm(10)
b1 <- 1:10
b <- rbind(b0, b1)
Y <- matrix(NA, ncol = 10, nrow = 40)

for (i in 1:40) {
 for (j in 1:10) {
   temp_mean <- exp(X[i,,drop = FALSE] \%*\% b[,j,drop = FALSE] + z[i])
   Y[i,j] <- rpois(1, lambda = temp_mean)
 }
}

res <- fit_alg1(Y = Y, X = X, constraint_fn = function(x) {median(x)})

}
