% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_null_mle.R
\name{fit_null_mle}
\alias{fit_null_mle}
\title{Estimate under the null model
Estimate B and z parameters through block coordinate descent to maximize the unconstrained log likelihood function under a simple null hypothesis and given constraint.}
\usage{
fit_null_mle(
  formula_rhs = NULL,
  Y,
  X = NULL,
  covariate_data = NULL,
  B = NULL,
  constraint,
  constraint_cat = 1,
  subset_j = NULL,
  null_k = NULL,
  null_j = NULL,
  tolerance = 1e-10,
  tolerance_nr = 1e-10,
  use_tolerance = TRUE,
  maxit = 1000,
  maxit_nr = 1000,
  ncores = NULL,
  solve_option
)
}
\arguments{
\item{formula_rhs}{The right hand side of a formula specifying which covariates to include in the model, must be used with the \code{covariate_data} parameter or replaced by the \code{X} parameter.}

\item{Y}{An outcome matrix with n rows (for samples) and J columns (for categories) containing abundance data.}

\item{X}{Design matrix with n rows (for samples) and p columns (for covariates).}

\item{covariate_data}{A data frame including all covariates given in \code{formula_rhs}, can be replaced by design matrix \code{X}.}

\item{B}{Optional initial parameter estimate for B matrix.}

\item{constraint}{Should be "scc" for single category constraint, "mc" for mean constraint, or "msc" for mean
over a subset constraint.}

\item{constraint_cat}{Category to constrain coefficients to equal the negative sum of all other categories.}

\item{subset_j}{Indices of categories to include in constraint for the mean over a subset constraint.}

\item{null_k}{Coefficient of the covariate in design matrix that is set to \code{0} under the null hypothesis.}

\item{null_j}{Category for which covariate \code{k} is set to \code{0} under the null hypothesis.}

\item{tolerance}{The tolerance used to stop the algorithm when log likelihood values are within \code{tolerance} of each other.}

\item{tolerance_nr}{The tolerance used to stop the Newton-Raphon method.}

\item{use_tolerance}{If \code{FALSE}, will run until \code{maxit} regardless of convergence.}

\item{maxit}{The maximum number of iterations of the coordinate descent algorithm.}

\item{maxit_nr}{The maximum number of iterations of the Newton-Raphson algorithm within the coordinate descent.}

\item{ncores}{The desired number of cores to optimize block of B parameters in parallel. If not provided, an appropriate number will be chosen for your machine.}

\item{solve_option}{}
}
\value{
A list including values of the log likelihood, the B matrix, and the z vector at each iteration.
}
\description{
Estimate under the null model
Estimate B and z parameters through block coordinate descent to maximize the unconstrained log likelihood function under a simple null hypothesis and given constraint.
}
\examples{
J <- 50
n <- 40
X <- cbind(1, rep(c(0, 1), each = n/2))
z <- rnorm(n) + 8
b0 <- rnorm(J)
b1 <- rnorm(J)
b <- rbind(b0, b1)
Y <- matrix(NA, ncol = J, nrow = n)

for (i in 1:n) {
 for (j in 1:J) {
   temp_mean <- exp(X[i, , drop = FALSE] \%*\% b[, j, drop = FALSE] + z[i])
   Y[i,j] <- rpois(1, lambda = temp_mean)
 }
}

null_mle <- fit_null_mle(Y = Y, X = X, ncores = 2, null_k = 2, null_j = 2,
                          constraint = "scc", solve_option = "solve")

}
