% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_A_prime.R
\name{generate_A_prime}
\alias{generate_A_prime}
\title{Generate A prime
Create expanded design matrix A prime.}
\usage{
generate_A_prime(X_prime, J)
}
\arguments{
\item{X_prime}{Expanded matrix of X values, nJ rows by pJ columns}

\item{J}{The number of KOs.}
}
\value{
A_prime, a matrix with \code{nJ} rows and \code{pJ + n} columns.
}
\description{
Generate A prime
Create expanded design matrix A prime.
}
