% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_B_tilde.R
\name{generate_B_tilde}
\alias{generate_B_tilde}
\title{Generate B tilde
Compress B matrix into vector B tilde.}
\usage{
generate_B_tilde(B)
}
\arguments{
\item{B}{Matrix of B values, p rows by J columns}
}
\value{
B_tilde, a vector of length p*J
}
\description{
Generate B tilde
Compress B matrix into vector B tilde.
}
