% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_theta.R
\name{generate_theta}
\alias{generate_theta}
\title{Generate theta
Put all B and z parameters together into one vector.}
\usage{
generate_theta(B_vec, z)
}
\arguments{
\item{B_vec}{vector of B values of length p*J.}

\item{z}{vector of z values of length n.}
}
\value{
theta, a vector of length p*J + n
}
\description{
Generate theta
Put all B and z parameters together into one vector.
}
