% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_W.R
\name{generate_W}
\alias{generate_W}
\title{Generate W
Create diagonal matrix with expected value of Y tilde on diagonals.}
\usage{
generate_W(X_tilde, theta)
}
\arguments{
\item{X_tilde}{An expanded version of the design matrix X.}

\item{theta}{A vector of all B and z parameters}
}
\value{
W, a diagonal matrix with \code{nJ} rows and columns.
}
\description{
Generate W
Create diagonal matrix with expected value of Y tilde on diagonals.
}
