% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_W_cstr.R
\name{generate_W_cstr}
\alias{generate_W_cstr}
\title{Variance of score
Calculate the empirical variance of the score for the constrained likelihood.}
\usage{
generate_W_cstr(W, X, B, z, constraint, constraint_cat, subset_j)
}
\arguments{
\item{W}{W matrix from unconstrained likelihood, diagonal matrix with entry i, j giving expected Y_ij based on mean model.}

\item{X}{Design matrix with n rows (for samples) and p columns (for covariates).}

\item{B}{Beta parameter matrix with p rows (for covariates) and J columns (for categories).}

\item{z}{z parameter vector of length n.}

\item{constraint}{Should be "scc" for single category constraint, "mc" for mean constraint, or "msc" for mean
over a subset constraint.}

\item{constraint_cat}{Category to constrain coefficients to equal the negative sum of all other categories.}

\item{subset_j}{Indices of categories to include in constraint for the mean over a subset constraint.}
}
\value{
A matrix giving the empirical variance of the score vector.
}
\description{
Variance of score
Calculate the empirical variance of the score for the constrained likelihood.
}
