% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_X_prime.R
\name{generate_X_prime}
\alias{generate_X_prime}
\title{Generate X prime
Create expanded design matrix X prime.}
\usage{
generate_X_prime(X, J)
}
\arguments{
\item{X}{Matrix of X values, n rows by p columns}

\item{J}{The number of KOs.}
}
\value{
X_prime, a matrix with \code{nJ} rows and \code{pJ} columns.
}
\description{
Generate X prime
Create expanded design matrix X prime.
}
