% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_X_star.R
\name{generate_X_star}
\alias{generate_X_star}
\title{Generate X star
Create expanded design matrix X star (the first pJ columns of X_tilde).}
\usage{
generate_X_star(X, J, transpose = FALSE)
}
\arguments{
\item{X}{Matrix of X values, n rows by p columns}

\item{J}{The number of KOs.}

\item{transpose}{If true, will generate the transpose of X_star. Set to \code{FALSE}.}
}
\value{
X_star, a matrix with \code{nJ} rows and \code{pJ} columns.
}
\description{
Generate X star
Create expanded design matrix X star (the first pJ columns of X_tilde).
}
