% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_Y_plus.R
\name{generate_Y_plus}
\alias{generate_Y_plus}
\title{Generate Y plus
Generate an augmented outcome matrix to use with Firth penalized likelihood.}
\usage{
generate_Y_plus(Y_tilde_plus, J)
}
\arguments{
\item{Y_tilde_plus}{An augmented outcome vector of length \code{nJ}.}

\item{J}{The number of KOs.}
}
\value{
Y_plus, the augmented outcome matrix.
}
\description{
Generate Y plus
Generate an augmented outcome matrix to use with Firth penalized likelihood.
}
