% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_Y_tilde.R
\name{generate_Y_tilde}
\alias{generate_Y_tilde}
\title{Generate Y tilde
Compress Y matrix into vector Y tilde.}
\usage{
generate_Y_tilde(Y)
}
\arguments{
\item{Y}{Matrix of Y values, n rows by J columns}
}
\value{
Y_tilde, a vector of length n*J
}
\description{
Generate Y tilde
Compress Y matrix into vector Y tilde.
}
