% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_Y_tilde_plus.R
\name{generate_Y_tilde_plus}
\alias{generate_Y_tilde_plus}
\title{Generate Y tilde plus
Generate an augmented outcome vector to use with Firth penalized likelihood.}
\usage{
generate_Y_tilde_plus(Y_tilde, X_tilde, W)
}
\arguments{
\item{Y_tilde}{An outcome vector of length \code{nJ}.}

\item{X_tilde}{An expanded version of the design matrix.}

\item{W}{A diagonal matrix with expected values of Y tilde on the diagonal.}
}
\value{
Y_tilde_plus, the augmented outcome vector.
}
\description{
Generate Y tilde plus
Generate an augmented outcome vector to use with Firth penalized likelihood.
}
