% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_augment_diag.R
\name{get_augment_diag}
\alias{get_augment_diag}
\title{Generate values to augment Y
Generate the diagonal values of the augmentation matrix, a helper function for \code{generate_Y_tilde_plus}}
\usage{
get_augment_diag(i)
}
\arguments{
\item{i}{The index of the diagonal.}
}
\value{
The value of the augmentation matrix for the ith diagonal.
}
\description{
Generate values to augment Y
Generate the diagonal values of the augmentation matrix, a helper function for \code{generate_Y_tilde_plus}
}
