% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_hat.R
\name{get_hat}
\alias{get_hat}
\title{Get hat matrix
Calculate diagonals of hat matrix for Poisson log linear model}
\usage{
get_hat(X, B, z, cc = NULL)
}
\arguments{
\item{X}{Design matrix with n rows (for samples) and p columns (for covariates), should have a leading intercept column of \code{1}s.}

\item{B}{Beta matrix}

\item{z}{z vector}

\item{cc}{Constraint category}
}
\value{
The diagonals of the hat matrix for the Poisson log linear model with design matrix X, parameters B, offsets z, and cc category used as the baseline category.
}
\description{
Get hat matrix
Calculate diagonals of hat matrix for Poisson log linear model
}
\examples{
n <- 50
J <- 10
cc <- 10
#X <- cbind(1, rnorm(n))
X <- cbind(1, rep(0:1, n/2))
b0 <- c(rnorm(J - 1), 0)
b1 <- c(rnorm(J - 1), 0)
B <- rbind(b0, b1)
z_init <- rnorm(n) + 8
Y <- matrix(NA, ncol = J, nrow = n)

for (i in 1:n) {
 for (j in 1:J) {
   temp_mean <- exp(X[i, , drop = FALSE] \%*\% B[, j, drop = FALSE] + z_init[i])
   Y[i,j] <- rpois(1, lambda = temp_mean)
 }
}
z <- update_z(Y, X, B)
get_hat(X, B, z, cc)

}
