% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_theta_ind.R
\name{get_theta_ind}
\alias{get_theta_ind}
\title{Switch indices
Go from index in Beta matrix to theta vector.}
\usage{
get_theta_ind(j, k, p)
}
\arguments{
\item{j}{Category or vector of categories.}

\item{k}{Covariate index or vector of covariate indices.}

\item{p}{Number of covariates.}
}
\value{
The index of Beta_k^j in the theta vector.
}
\description{
Switch indices
Go from index in Beta matrix to theta vector.
}
