% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/null_score_var.R
\name{null_score_var}
\alias{null_score_var}
\title{Variance of score
Calculate the empirical variance of the score for the constrained likelihood.}
\usage{
null_score_var(Y, X, B, z, constraint, constraint_cat)
}
\arguments{
\item{Y}{An outcome matrix with n rows (for samples) and J columns (for KOs) containing coverage data.}

\item{X}{Design matrix with n rows (for samples) and p columns (for covariates), should have a leading intercept column of \code{1}s, can be replaced by \code{formula_rhs} and \code{covariate_data}.}

\item{B}{B parameter matrix.}

\item{z}{z parameter vector.}

\item{constraint}{Should be "scc" for single category constraint or "mc" for mean constraint.}

\item{constraint_cat}{Category to constrain coefficients (for single category the B parameters for this
category will equal 0, for mean the B parameters for this category will equal the negative sum of the others.}
}
\value{
The empirical variance of the score.
}
\description{
Variance of score
Calculate the empirical variance of the score for the constrained likelihood.
}
\examples{
X <- cbind(1, rep(c(0, 1), each = 20))
z <- rnorm(40) + 8
b0 <- rnorm(10)
b1 <- 1:10
b <- rbind(b0, b1)
Y <- matrix(NA, ncol = 10, nrow = 40)

for (i in 1:40) {
 for (j in 1:10) {
   temp_mean <- exp(X[i, , drop = FALSE] \%*\% b[, j, drop = FALSE] + z[i])
   Y[i,j] <- rpois(1, lambda = temp_mean)
 }
}

res <- null_score_var(Y = Y, X = X, B = b, z = z, 
                      constraint = "scc", constraint_cat = 1)

}
