% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_score_test.R
\name{run_score_test}
\alias{run_score_test}
\title{Run a robust score test
Calculate test statistic and p-value for robust score test under given constraint.}
\usage{
run_score_test(
  formula_rhs = NULL,
  Y,
  X = NULL,
  covariate_data = NULL,
  B = NULL,
  constraint,
  constraint_cat = 1,
  null_k,
  null_j,
  robust = TRUE,
  tolerance = 1e-10,
  tolerance_nr = 1e-10,
  use_tolerance = TRUE,
  maxit = 1000,
  maxit_nr = 1000,
  ncores = NULL
)
}
\arguments{
\item{formula_rhs}{The right hand side of a formula specifying which covariates to include in the model, must be used with the \code{covariate_data} parameter or replaced by the \code{X} parameter.}

\item{Y}{An outcome matrix with n rows (for samples) and J columns (for categories) containing abundance data.}

\item{X}{Design matrix with n rows (for samples) and p columns (for covariates).}

\item{covariate_data}{A data frame including all covariates given in \code{formula_rhs}, can be replaced by design matrix \code{X}.}

\item{B}{Optional initial parameter estimate for B matrix.}

\item{constraint}{Should be "scc" for single category constraint, "mc" for mean constraint, or "msc" for mean
over a subset constraint.}

\item{constraint_cat}{Category to constrain coefficients to equal the negative sum of all other categories.}

\item{null_k}{Coefficient of the covariate in design matrix that is set to \code{0} under the null hypothesis.}

\item{null_j}{Category for which covariate \code{k} is set to \code{0} under the null hypothesis.}

\item{robust}{If TRUE runs a robust score test, if FALSE runs a model-based score test.}

\item{tolerance}{The tolerance used to stop the algorithm when log likelihood values are within \code{tolerance} of each other.}

\item{tolerance_nr}{The tolerance used to stop the Newton-Raphon method.}

\item{use_tolerance}{If \code{FALSE}, will run until \code{maxit} regardless of convergence.}

\item{maxit}{The maximum number of iterations of the coordinate descent algorithm.}

\item{maxit_nr}{The maximum number of iterations of the Newton-Raphson algorithm within the coordinate descent.}

\item{ncores}{The desired number of cores to optimize block of B parameters in parallel. If not provided, an appropriate number will be chosen for your machine.}
}
\value{
A list including values of the log likelihood, the B matrix, and the z vector at each iteration.
}
\description{
Run a robust score test
Calculate test statistic and p-value for robust score test under given constraint.
}
\examples{
X <- cbind(1, rep(c(0, 1), each = 20))
z <- rnorm(40) + 8
b0 <- rnorm(10)
b1 <- 1:10
b <- rbind(b0, b1)
Y <- matrix(NA, ncol = 10, nrow = 40)

for (i in 1:40) {
 for (j in 1:10) {
   temp_mean <- exp(X[i, , drop = FALSE] \%*\% b[, j, drop = FALSE] + z[i])
   Y[i,j] <- rpois(1, lambda = temp_mean)
 }
}

res <- run_score_test(Y = Y, X = X, ncores = 2, null_k = 2, null_j = 2,
                      constraint = "scc")

}
