% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_Bj_list.R
\name{update_Bj_list}
\alias{update_Bj_list}
\title{Update Bj values
Use Poisson regression to update the Bj block in the block coordinate descent algorithm.}
\usage{
update_Bj_list(theta)
}
\arguments{
\item{theta}{A list containing all necessary parameters for this function. This list includes \code{j}, the current KO to be updated, \code{Y}, the outcome matrix, \code{X}, the design matrix, \code{B}, the current value of the B matrix, and \code{z}, the current value of the z vector.}
}
\value{
A vector of new values of B for KO j, the coefficients from fitting a glm.
}
\description{
Update Bj values
Use Poisson regression to update the Bj block in the block coordinate descent algorithm.
}
