% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_lambda.R
\name{update_lambda}
\alias{update_lambda}
\title{Update lambda values for constraint
Update lambda values used for constrained estimation under the null hypothesis.}
\usage{
update_lambda(Y, X, B, z, j)
}
\arguments{
\item{Y}{An outcome matrix with n rows (for samples) and J columns (for KOs) containing coverage data.}

\item{X}{Design matrix with n rows (for samples) and p columns (for covariates).}

\item{B}{Beta parameter matrix with p rows (for covariates) and J columns (for KOs).}

\item{z}{z parameter vector of length n.}

\item{j}{Category used for constraint.}
}
\value{
A vector of length p of lambda values.
}
\description{
Update lambda values for constraint
Update lambda values used for constrained estimation under the null hypothesis.
}
