% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_lambda_null.R
\name{update_lambda_null}
\alias{update_lambda_null}
\title{Update lambda values for null hypothesis
Update lambda values used for constrained estimation under the null hypothesis.}
\usage{
update_lambda_null(Y, X, B, z, null_j, null_k)
}
\arguments{
\item{Y}{An outcome matrix with n rows (for samples) and J columns (for KOs) containing coverage data.}

\item{X}{Design matrix with n rows (for samples) and p columns (for covariates).}

\item{B}{Beta parameter matrix with p rows (for covariates) and J columns (for KOs).}

\item{z}{z parameter vector of length n.}

\item{null_j}{Category tested in null hypothesis.}

\item{null_k}{Coefficient index tested in null hypothesis.}
}
\value{
A vector of length p of lambda values.
}
\description{
Update lambda values for null hypothesis
Update lambda values used for constrained estimation under the null hypothesis.
}
