% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_z.R
\name{update_z}
\alias{update_z}
\title{Update z values
Use the closed form solution to the score equation for z to update the block of z parameters in the block coordinate descent algorithm.}
\usage{
update_z(Y, X, B)
}
\arguments{
\item{Y}{An outcome matrix with n rows (for samples) and J columns (for KOs) containing coverage data.}

\item{X}{Design matrix with n rows (for samples) and p columns (for covariates).}

\item{B}{Beta parameter matrix with p rows (for covariates) and J columns (for KOs).}
}
\value{
A vector of new values of z.
}
\description{
Update z values
Use the closed form solution to the score equation for z to update the block of z parameters in the block coordinate descent algorithm.
}
