% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/var_single_score.R
\name{var_single_score}
\alias{var_single_score}
\title{Estimate variance of single score value
Calculate the empirical variance of the score for the parameter tested.}
\usage{
var_single_score(null_j, null_k, upd_ind, p, D, info)
}
\arguments{
\item{null_j}{Category for which covariate \code{k} is set to \code{0} under the null hypothesis.}

\item{null_k}{Coefficient of the covariate in design matrix that is set to \code{0} under the null hypothesis.}

\item{upd_ind}{Indices of parameter vector, not including parameters from constraint row or parameter implicated in null hypothesis.}

\item{p}{Number of columns in design matrix.}

\item{D}{Estimated empirical variance of score.}

\item{info}{Information matrix.}
}
\value{
The empirical variance of the score for a single value.
}
\description{
Estimate variance of single score value
Calculate the empirical variance of the score for the parameter tested.
}
