% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_alg1.R
\name{fit_alg1}
\alias{fit_alg1}
\title{Fit Algorithm 1
Estimate B and z parameters through block coordinate descent to maximize the log likelihood function, details given in Algorithm 1.}
\usage{
fit_alg1(
  formula_rhs = NULL,
  Y,
  X = NULL,
  covariate_data = NULL,
  B = NULL,
  tolerance = 0.1,
  maxit = 100,
  constraint_fn,
  maxit_glm = 100,
  ncores = NULL
)
}
\arguments{
\item{formula_rhs}{The right hand side of a formula specifying which covariates to include in the model, must be used with the \code{covariate_data} parameter or replaced by the \code{X} parameter.}

\item{Y}{An outcome matrix with n rows (for samples) and J columns (for KOs) containing coverage data.}

\item{X}{Design matrix with n rows (for samples) and p columns (for covariates), should have a leading intercept column of \code{1}s, can be replaced by \code{formula_rhs} and \code{covariate_data}.}

\item{covariate_data}{A data frame including all covariates given in \code{formula_rhs}, can be replaced by design matrix \code{X}.}

\item{B}{Optional initial parameter estimate for B matrix}

\item{tolerance}{The tolerance used to stop the algorithm when log likelihood values are within \code{tolerance} of each other.}

\item{maxit}{The maximum number of iterations of the coordinate descent algorithm.}

\item{constraint_fn}{A constraint function to make the B matrix identifiable.}

\item{maxit_glm}{The maximum number of iterations when running the glm to update the block of Bj parameters in the coordinate descent algorithm.}

\item{ncores}{The desired number of cores to optimize block of B parameters in parallel. If not provided, an appropriate number will be chosen for your machine.}
}
\value{
A list including values of the log likelihood, the B matrix, and the z vector at each iteration.
}
\description{
Fit Algorithm 1
Estimate B and z parameters through block coordinate descent to maximize the log likelihood function, details given in Algorithm 1.
}
\examples{
X <- cbind(1, rep(c(0, 1), each = 20))
z <- rnorm(40) + 8
b0 <- rnorm(10)
b1 <- 1:10
b <- rbind(b0, b1)
Y <- matrix(NA, ncol = 10, nrow = 40)

for (i in 1:40) {
 for (j in 1:10) {
   temp_mean <- exp(X[i, , drop = FALSE] \%*\% b[, j, drop = FALSE] + z[i])
   Y[i,j] <- rpois(1, lambda = temp_mean)
 }
}

res <- fit_alg1(Y = Y, X = X, constraint_fn = function(x) {median(x)}, ncores = 2)

}
